Create SagasManager
`SagasManager` is a central repository for all sagas in the application.

`SagasManager` has two methods: `addSagaToRoot` and `getRootSaga`.
`addSagaToRoot` takes a saga and any additional arguments for that saga
and adds it to its internal array as an array. `getRootSaga` returns a
generator that yields a map of each saga/argument array in the internal
array as a fork.  A new `SagasManager` instance is exported as a
singleton from the file.

Sagas for each module will be combined into a single watcher function
which yields an array containing the take-type call for each saga. That
single watcher function is exported from the module as `sagas` from the
module. That watcher function is then imported in a central place and
added to the `SagasManager` singleton using its `addSagaToRoot` method.

`SagasManager` is then imported into `index.js`. The configured `store`
will have a `runSaga` method due to the middleware imported when
creating the store. The `store.runSaga` method is called with
`sagasManager.getRootSaga()` is the argument. This will allow all of the
combined watchers to subscribe to the store to watch for the given
action type calls.