Create formatResponse actions
formatResponse takes normalized data and alters the information. It
returns the entities key of the normalized response without changing it
but recreates the result key.

The result key is a Map whose keys are the different types of entities
and whose values are lists of the ids of those entities.