Rewrite Login/LoginForm components
The LoginForm component was exported with its reduxForm container and
the Login component was a Redux container around that combination. This
didn't work for testing as we need to test the rendered component itself
and its containers separately.

By exporting the base presentational component as LoginForm and the
fully connected container as default, we are able to test presentation
and connections separately.